# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API Quản trị Nhân sự HUIT
  description: Tài liệu RESTful API cho hệ thống Quản trị nhân sự HUIT (ERP Backend) do Hóa Coder phát triển.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Dev server kết nối Database Local/Docker

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Department:
      type: object
      properties:
        MA_PHG:
          type: integer
        TEN_PB:
          type: string
        NGAY_THANH_LAP:
          type: string
          format: date-time
        MA_TRUONG_PHG:
          type: string

    Employee:
      type: object
      properties:
        HO_TEN:
          type: string
        EMAIL:
          type: string
        CHUC_VU:
          type: string
        LUONG_CB:
          type: number
        MA_PHG:
          type: integer
        NGAY_SINH:
          type: string
          format: date
        GIOI_TINH:
          type: string
        DIA_CHI:
          type: string

    Project:
      type: object
      properties:
        TEN_DA:
          type: string
        MO_TA:
          type: string
        NGAY_BAT_DAU:
          type: string
          format: date
        NGAY_KET_THUC:
          type: string
          format: date
        TRANG_THAI:
          type: string

  responses:
    SuccessCreate:
      description: Thêm mới thành công
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
    SuccessUpdate:
      description: Cập nhật thành công
    ErrorResponse:
      description: Dữ liệu không hợp lệ hoặc thiếu quyền
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string

security:
  - bearerAuth: []

paths:
  # =========================================================================
  # AUTH - XÁC THỰC
  # =========================================================================
  /auth/register:
    post:
      summary: Đăng ký tài khoản nhân viên mới (Chờ OTP)
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, hoten]
              properties:
                email:
                  type: string
                password:
                  type: string
                hoten:
                  type: string
      responses:
        "201":
          $ref: "#/components/responses/SuccessCreate"

  /auth/verify:
    post:
      summary: Xác thực OTP và kích hoạt tài khoản
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, otpCode]
              properties:
                email:
                  type: string
                otpCode:
                  type: string
      responses:
        "200":
          $ref: "#/components/responses/SuccessUpdate"

  /auth/login:
    post:
      summary: Đăng nhập vào hệ thống
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                  user:
                    type: object

  /auth/change-password:
    post:
      summary: Đổi mật khẩu
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        "200":
          $ref: "#/components/responses/SuccessUpdate"

  # =========================================================================
  # NHÂN VIÊN
  # =========================================================================
  /employees:
    get:
      summary: Lấy danh sách toàn bộ nhân viên (Có phân trang)
      tags: [Employees]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 10
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Trả về danh sách
    post:
      summary: HR/Admin tạo mới nhân sự
      tags: [Employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
      responses:
        "201":
          $ref: "#/components/responses/SuccessCreate"

  /employees/{id}:
    get:
      summary: Xem chi tiết hồ sơ 1 nhân viên
      tags: [Employees]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trả về thông tin NV
    put:
      summary: Cập nhật thông tin nhân viên
      tags: [Employees]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
      responses:
        "200":
          $ref: "#/components/responses/SuccessUpdate"

  # =========================================================================
  # PHÒNG BAN
  # =========================================================================
  /departments:
    get:
      summary: Lấy danh sách phòng ban
      tags: [Departments]
      responses:
        "200":
          description: Danh sách phòng ban
    post:
      summary: Thêm phòng ban
      tags: [Departments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [maphg, tenpb]
              properties:
                maphg:
                  type: integer
                tenpb:
                  type: string
                matruongphg:
                  type: string
      responses:
        "201":
          $ref: "#/components/responses/SuccessCreate"

  /departments/{id}:
    get:
      summary: Xem phòng ban và danh sách nhân sĩ trong phòng đó
      tags: [Departments]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chi tiết
    put:
      summary: Sửa đổi (Đổi tên, Trưởng phòng)
      tags: [Departments]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tenpb:
                  type: string
                matruongphg:
                  type: string
      responses:
        "200":
          $ref: "#/components/responses/SuccessUpdate"

  # =========================================================================
  # DỰ ÁN
  # =========================================================================
  /projects:
    get:
      summary: Lấy danh sách toàn bộ dự án
      tags: [Projects]
      responses:
        "200":
          description: Trả về mảng dự án
    post:
      summary: Tạo mới dự án
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Project"
      responses:
        "201":
          $ref: "#/components/responses/SuccessCreate"

  /projects/{id}:
    get:
      summary: Lấy chi tiết dự án (Kèm các thành viên)
      tags: [Projects]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chi tiết Project
    put:
      summary: Cập nhật trạng thái dự án
      tags: [Projects]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Project"
      responses:
        "200":
          $ref: "#/components/responses/SuccessUpdate"

  /projects/employee/{id}:
    get:
      summary: Lấy danh sách các dự án mà 1 nhân viên tham gia
      tags: [Projects]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: Mã nhân viên
      responses:
        "200":
          description: Danh sách dự án cá nhân

  /projects/{id}/members:
    post:
      summary: Thêm 1 nhân sự vào dự án
      tags: [Projects]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manv, vaitroduan]
              properties:
                manv:
                  type: string
                vaitroduan:
                  type: string
      responses:
        "201":
          $ref: "#/components/responses/SuccessCreate"

  # =========================================================================
  # TÍNH LƯƠNG & CHẤM CÔNG
  # =========================================================================
  /payroll/{year}/{month}:
    get:
      summary: Xem bảng lương toàn công ty của tháng
      tags: [Payroll]
      parameters:
        - in: path
          name: year
          required: true
          schema:
            type: integer
        - in: path
          name: month
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trả về danh sách lương của tháng

  /payroll/employee/{id}:
    get:
      summary: Xem phiếu lương cá nhân theo tháng/năm
      tags: [Payroll]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: month
          required: false
          schema:
            type: integer
        - in: query
          name: year
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Trả về phiếu lương của nhân viên

  /payroll/salary/{id}:
    put:
      summary: Cập nhật dữ liệu bảng lương
      tags: [Payroll]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                SO_NGAY_CONG_THUC_TE:
                  type: number
                PHU_CAP:
                  type: number
                THUONG:
                  type: number
                KHAU_TRU_BHXH:
                  type: number
      responses:
        "200":
          $ref: "#/components/responses/SuccessUpdate"

  /payroll/check-in:
    post:
      summary: Chấm công vào (Check-in)
      tags: [Attendance]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [MA_NV]
              properties:
                MA_NV:
                  type: string
      responses:
        "200":
          description: Check-in thành công

  /payroll/check-out:
    post:
      summary: Chấm công ra (Check-out)
      tags: [Attendance]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [MA_NV]
              properties:
                MA_NV:
                  type: string
      responses:
        "200":
          description: Check-out thành công